BIG UPDATE - HEADERS 12/07/2017
Divide into diff source files
